Updated Report to reflect send report via email function, added new classes for row and sheet, created test file.
